Add testbench circlebuffer & removed dumpRingBuffer
Changed how dumpRingBuffer worked and gave it a proper name.
Rather than triggering on a single clock and reading out the entire
content, it now reads out values one by one when given the clock signal
is high.

NOTE: this new implementation will *never* reset it's state and is meant
for testing only