Fix packet parsing
The system beforehand assumed the bytes iterator would send 'complete'
packets. Which is a foolish assumption to make. This remedies that by
adding in support for parsing incomplete packets.

Although; there's no proper error handling as of now, if a packet is
incomplete the system will simply freak out until it magically resolves
itself.

I have an idea for that though; I could add a preamble, and then it
fails it will try to clear the buffer up until the preamble. Then from
there it can try again.