Add seperation of preamble and packet type
Split the preamble and packet type into it's own seperate circuit.

This allows individual packet types to have to carry less data which
isn't relevant to them. It also allows for a better type semantics as
there is now a distinction between 'work in progress' packets and 'ready
to send' packets.