Refactor polyvariadic function
Authored by: @NeuroCorgi

I simply copied his code in this commit, this is because cherry picking
across PR's in Github is stupid and messy. And merging the original PR
would take very long due to it being out of date.

It now no longer requires an empty tuple to end the polyvariadic
function. Cleaning up the API a little more.