Wishbone (& etherbone) integration (#56)
* Upgraded clash-compiler and added clash-cores

As clash-cores has been removed from clash-compiler, I had to add it
seperately

* Add df2ps

* Add unsafeCSignalToDf and mergeBytePackets

Utility functions for handling PS.

* Added etherboneDfPacketizer

* Added a wishbone interface to ringBuffer

* Add cli trigger configuration template generation

* Add barebones etherbone abstraction layer in cli

* Fix etherboneDfPacketizer breaking when multiple packets being sent

* Add memory map description

* Add cli etherbone packetizer

* Renamed to_record to to_records as it returns a vector

* Add small etherbone communication demo

* Add cli Etherbone packet generation

* Fix contradiction in ilaCore

If 'freeze' is true, it should probably freeze the buffer, not do the
inverse.

* Add wishbone utils

A small utility library to make working with wishbone addresses a bit
easier

* Add addrMuxWO

* Add ILA WB buffer manager

* Removed addrMuxWO

addrMux already only works on write cycles. So this did nothing.

* Fix wishbone utilities not working in simulation

In simulation these functions break due to `undefined` addresses.
Changing the order around fixes it.

* Removed memory map comment

* Add ilaWb component

Adds a wishbone interface to the ILA. Can be used to alter the behaviour
of the ILA during runtime.

* Add ilaUart component

A specialized component to instantiate an ILA working with UART.

* Fix cli read commands

* Add toplevel using ilaUart

* Split off ila from ilaUart

Splits the etherbone ILA off from the ilaUart component. Makes the
entire system a little bit more modular.

* Remove old components

A big old commit removing a bunch of now unused components. Mostly the
old packet communication stuff.

* BROKEN - Refactor code codebase part 1

Refactors significant parts of the codebase to be structured in a more
logical way. Also removes the code used for the custom packetting format
for an etherbone messaging format.

* BROKEN - Moved version and sizes to const

* BROKEN - Perform returns only relevant bytes

* BROKEN - Move signal parsing to SignalCluster definition

* BROKEN - Add RegisterOutput and modular etherbone output parsing

* BROKEN - Make IlaRegister properly input logical indices rather than raw

* BROKEN - Add perform_register_operation

Adds function simplifying write/read transactions to the ila. Rather
than having to call the individual methods on the enums and records, you
can give this function a register and it will do the work for you.

* BROKEN - Removed vcd and analysis modes from CLI

They no longer make sense to keep around with the new etherbone
messaging format

* Unbroke Rust CLI

Final changes to make the CLI not be broken. Due to the new api, it
can't capture signals yet though.

* Add (back) signal capture

* Fix incorrect addresses in CLI

* Add Rust CLI probe for hash

Checks if the hash stored on the ILA matches the one in the given
configuration file. Used to ensure that we're talking to the device we
want to talk to and not some old or outdated one.

* Fix cabal pointing to local directory rather online clash-cores

This was done for debugging and I forgot to change it back

Also we can now point to main clash-cores since etherboneC has been
fully merged.

* Stole signal field selectors from Bittide

Copied the file straight from bittide-hardware. Only difference being
the module name which I changed so I wouldn't need to nest a lot of
directories.

* Add non signal wbCycle check

* Refactored ilaWb

Rather than using a lot of different signals to store each of the
registers seperately. Now there's a proper solution using a record and
moore machine.

* Implement haskell feedback PR

* Add cargo clippy & apply clippy changes