Fix broken predicate configuration & proper VCD timescale
The implicit clock cycle was editable by the user via the tui. This is
unwanted. Removing this however introduced a few other bugs relating to
predicates.

This implicit clock is now used by the VCD export to accurately display
the duration of a clock signal. The duration is always displayed in
nanoseconds.