Merge library (#73)
* Upgraded clash-compiler and added clash-cores

As clash-cores has been removed from clash-compiler, I had to add it
seperately

* Add df2ps

* Add unsafeCSignalToDf and mergeBytePackets

Utility functions for handling PS.

* Added etherboneDfPacketizer

* Added a wishbone interface to ringBuffer

* Add cli trigger configuration template generation

* Add barebones etherbone abstraction layer in cli

* Fix etherboneDfPacketizer breaking when multiple packets being sent

* Add memory map description

* Add cli etherbone packetizer

* Renamed to_record to to_records as it returns a vector

* Add small etherbone communication demo

* Add cli Etherbone packet generation

* Fix contradiction in ilaCore

If 'freeze' is true, it should probably freeze the buffer, not do the
inverse.

* Add wishbone utils

A small utility library to make working with wishbone addresses a bit
easier

* Add addrMuxWO

* Add ILA WB buffer manager

* Removed addrMuxWO

addrMux already only works on write cycles. So this did nothing.

* Fix wishbone utilities not working in simulation

In simulation these functions break due to `undefined` addresses.
Changing the order around fixes it.

* Removed memory map comment

* Add ilaWb component

Adds a wishbone interface to the ILA. Can be used to alter the behaviour
of the ILA during runtime.

* Add ilaUart component

A specialized component to instantiate an ILA working with UART.

* Fix cli read commands

* Add toplevel using ilaUart

* Split off ila from ilaUart

Splits the etherbone ILA off from the ilaUart component. Makes the
entire system a little bit more modular.

* Remove old components

A big old commit removing a bunch of now unused components. Mostly the
old packet communication stuff.

* BROKEN - Refactor code codebase part 1

Refactors significant parts of the codebase to be structured in a more
logical way. Also removes the code used for the custom packetting format
for an etherbone messaging format.

* BROKEN - Moved version and sizes to const

* BROKEN - Perform returns only relevant bytes

* BROKEN - Move signal parsing to SignalCluster definition

* BROKEN - Add RegisterOutput and modular etherbone output parsing

* BROKEN - Make IlaRegister properly input logical indices rather than raw

* BROKEN - Add perform_register_operation

Adds function simplifying write/read transactions to the ila. Rather
than having to call the individual methods on the enums and records, you
can give this function a register and it will do the work for you.

* BROKEN - Removed vcd and analysis modes from CLI

They no longer make sense to keep around with the new etherbone
messaging format

* Unbroke Rust CLI

Final changes to make the CLI not be broken. Due to the new api, it
can't capture signals yet though.

* Add (back) signal capture

* Fix incorrect addresses in CLI

* Add Rust CLI probe for hash

Checks if the hash stored on the ILA matches the one in the given
configuration file. Used to ensure that we're talking to the device we
want to talk to and not some old or outdated one.

* Fix cabal pointing to local directory rather online clash-cores

This was done for debugging and I forgot to change it back

Also we can now point to main clash-cores since etherboneC has been
fully merged.

* Stole signal field selectors from Bittide

Copied the file straight from bittide-hardware. Only difference being
the module name which I changed so I wouldn't need to nest a lot of
directories.

* Add non signal wbCycle check

* Refactored ilaWb

Rather than using a lot of different signals to store each of the
registers seperately. Now there's a proper solution using a record and
moore machine.

* Refactor ilaConfig

Changed the syntax of ilaConfig, it now requires a terminating case
again. This time that terminating case being an `WithIlaConfig` record.
This makes the syntax for creating an ILA really pretty.

Currently breaks compilation, due to me not having updated Blink.hs yet

* Fix Ila.hs not compiling with new IlaConfig changes

* Fix blackbox still expectiong dom argument

* Fix undeducable instance of ilaConfig & fix demo

Exposing the buffer depth and the final singal type in the polyvariadic
function does not go very well. Instead, keep them hidden and pattern
match on it when neccessary.

* Add predicate selection on the FPGA

You can now provide a predefined list of predicates to the ILA, and
depending on the `ILA Trigger Select` register value it will use that
predicate to apply to incoming samples. If the predicate holds true, the
system is triggered.

* Add predicate names to json file

* Add support for AND/OR'ing multiple predicates

* Implement FPGA support for multi-predicate selection

* Add rust UI utilities

* Split textinput to their own module

* Add predicate input UI

* Implement saving and loading of predicate configuration

* Add trigger state to TUI

* Fix incorrect bitordering

The ila-cli interpreted the order of the bits in the wrong order,
leading to broken results.

I don't quite know since when it was broken, since it did work. My
assumption is that in one of the polyvariadic function reworks the
ordering changed, whilst the cli didn't.

* Add two new default ilaPredicates

Added always true and always false predicates, may come in handy for
when adding support for runtime capture logic

* Rename orangecrab to clash-ila

* Updated README.md

* Ported latest changes from trigger-logic branch

* Posted latest changes from cli-arguments branch

* Squash merge implement-tests into library

* Include predicates documentation in README.md

* Updated gitignore

* Minor cleanups

* Add flake build dependencies

Rather than using cabal.project, use the flakes to manage dependencies

* Fix the evil clash team from sabotaging my project

THEY REMOVED DF.DATA AND NEVER TOLD ME
:(

* Removed dependency on libudev

* Add rust cli to programs and apps

* Fix clash not using nix dependencies

* Squash merge qol-fixes into library

* Fixed missing cli-macro dependency

After merging qol-fixes into library, it introduced a new local
dependency. buildRustPackage couldn't quite handle that, so I generated
Cargo.nix with `crate2nix` and imported it.

* Cleanup Blink.hs

Removed unneccesary functions from the demo, also increased the
bitwidths of the demo signals to demonstrate that it can measure bigger
values too.

* Fix inproper parsing of multi-word data

This commit fixes two issues:
 - When asking for buffer data, sending the wrong indices, leading to
wrong output
 - When parsing any multi-word data, the default into_vec() function
would start from MSB, however I need the data to be parsed from LSB.

* Make clash-protocols come as flake inputs

Rather than compiling clash-protocols locally, make it use flake inputs
for more effective caching (hopefully)

* Resolve merge conflict

* Update flake to use main clash-protocols repository

Also fixes a bug when merging that cabal.project didnt get removed.

* Update Nix flake infrastructure

* Clarified README

Fixed typos (wrong commit, incorrect keybind)
Added how to use Nix section
Added paragraph about masks

* Makefile rework & cleanup

Resolved issues regarding the makefile inside of the ila subdirectory
being quite inconsistent. Running `make` now should always succeed.

I also renamed the `Blink` project to `ClashIlaDemo`, moved `top/` to
`examples/` and removed unused `orangecrab/`, Pmod.hs and RGB.hs files.

* Higher default BAUD rate

* Add auto-sample (enabled by default)

Auto-sample automatically samples data from the ILA once the system
detected it is triggered

* Export Predicate and NamedPredicates

These types are quite useful for defining your own custom predicates