Restructure ILA buffer reads
Introduces two new registers inside of the ILA, selecting the specific
address and word index to read from the buffer.
This is contrast to how it was before, where memory was read from just
indexing addresses 0x3000_0000 to 0x3fff_ffff.

The main upside of the new approach is that this significantly improves
timing due to no longer needing an expensive `divMod`

The downside is that reading data is a LOT slower...