Bugfixes (#80)
* Add zero-bit mask check

If the trigger and mask are both zero, this would result in the
predicate firing true. This ensures that if the mask is zero, the
predicate will not be true (for the default predicates at least)

* Restructure ILA buffer reads

Introduces two new registers inside of the ILA, selecting the specific
address and word index to read from the buffer.
This is contrast to how it was before, where memory was read from just
indexing addresses 0x3000_0000 to 0x3fff_ffff.

The main upside of the new approach is that this significantly improves
timing due to no longer needing an expensive `divMod`

The downside is that reading data is a LOT slower...

* Remove BufferIndex buffer

Re-restructes how data is retrieved again. Removing the BufferIndex and
uses the consequentive memory read feature from Wishbone to streamline
reading data. Speeds up reading data *significantly*.

Also removes the WishboneUtils.hs file as it is no longer used anywhere